refactor: Centralize label socket event handling and update label state management
- Move label socket event listeners from LabelsSelector to TaskGroupWrapper
- Simplify label state initialization in tasks slice
- Update label filtering and selection logic
- Remove redundant socket connection checks
